Exercise completed
I pushed uncommitted, because I had the sub-folder problem. I deleted everything and subsequently reopened the project. The fact that the message on communications I read it at 6:10 pm, because I was not at home that I went to a bar to study. However I solved the problem, I also opened a ticket unaware that it was so late. Anyway solved the problem !!!